% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCC_seurat.r
\name{findMarkers_RCC}
\alias{findMarkers_RCC}
\title{Recursive Consensus Clustering}
\usage{
findMarkers_RCC(obj)
}
\arguments{
\item{config_file}{Input configuration file which specifies the absolute path for input files}
}
\value{
Cluster annotation file
}
\description{
Recursive Clustering of single-cell RNA-seq using Seurat as the base clustering tool.
}
