% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.r
\name{findMarkersAll}
\alias{findMarkersAll}
\title{Find markers from specific clusters}
\usage{
findMarkersAll(expr, info, fields)
}
\arguments{
\item{expr}{log2 CPM matrix with rows as genes and columns as cells}

\item{info}{meta data}

\item{fields}{meta data column names to be used as annotations for marker identification}
}
\value{
markers file
}
\description{
Find markers using normalize data in each cluster
}
